cleanups and bugfixes
