+++ src/reducers/dataSources/index.js
@@ -4,51 +4,19 @@ import {
   combineReducers,
   createSelector,
 } from '@reduxjs/toolkit';
-import actionTypes from './../../const/application-actions';
-
-const types = {
-  fetchAll: {
-    pending: `${actionTypes.DATASOURCE_FETCH_ALL}_PENDING`,
-    success: `${actionTypes.DATASOURCE_FETCH_ALL}_SUCCESS`,
-    fail: `${actionTypes.DATASOURCE_FETCH_ALL}_REJECT`,
-  },
-  fetchVersions: {
-    pending: `${actionTypes.DATASOURCE_FETCH_VERSIONS}_PENDING`,
-    success: `${actionTypes.DATASOURCE_FETCH_VERSIONS}_SUCCESS`,
-    fail: `${actionTypes.DATASOURCE_FETCH_VERSIONS}_REJECT`,
-  },
-  create: {
-    pending: `${actionTypes.DATASOURCE_CREATE}_PENDING`,
-    success: `${actionTypes.DATASOURCE_CREATE}_SUCCESS`,
-    fail: `${actionTypes.DATASOURCE_CREATE}_REJECT`,
-  },
-  fetch: {
-    pending: `${actionTypes.DATASOURCE_FETCH}_PENDING`,
-    success: `${actionTypes.DATASOURCE_FETCH}_SUCCESS`,
-    fail: `${actionTypes.DATASOURCE_FETCH}_REJECT`,
-  },
-  postVersion: {
-    pending: `${actionTypes.DATASOURCE_POST_VERSION}_PENDING`,
-    success: `${actionTypes.DATASOURCE_POST_VERSION}_SUCCESS`,
-    fail: `${actionTypes.DATASOURCE_POST_VERSION}_REJECT`,
-  },
-};
-
+import types from './actionTypes';
+import { reducer as versionsReducer } from './versions';
 /**
  * @typedef {import('./datasource.d').DataSource} DataSource
  * @typedef {import('@reduxjs/toolkit').EntityState<DataSource>} CollectionState
  * @typedef {import('./datasource').DataSourceEntry} DataSourceEntry
  * @typedef {import('./datasource').FetchStatus} FetchStatus
- * @typedef {import('./datasource').DataSourceVersion} DataSourceVersion
- * @typedef {{
- *   [name: string]: { status: FetchStatus; versions: DataSourceVersion[] };
- * }} VersionsCollection
+ * @typedef {import('./versions').VersionsState} VersionsState
  * @typedef {{
  *   collection: CollectionState;
  *   status: FetchStatus;
  *   error: string | null;
- *   activeVersions: { [name: string]: string };
- *   versions: VersionsCollection;
+ *   versions: VersionsState;
  * }} DataSourcesState
  * @typedef {{
  *   dataSources: DataSourcesState;
@@ -57,113 +25,6 @@ const types = {
 
 /** @type {import('@reduxjs/toolkit').EntityAdapter<DataSource>} */
 const entityAdapter = createEntityAdapter();
-
-// this is temporary!
-// in the future there will be a project config object setting which version is active
-// for now:
-// track the latest version for each dataSource
-const activeVersions = createSlice({
-  initialState: {},
-  name: 'dataSources/active-versions',
-  reducers: {},
-  extraReducers: {
-    /** @param {{ payload: DataSource[] }} action */
-    [types.fetchAll.success]: (state, action) =>
-      action.payload.reduce(
-        (acc, item) => ({ ...acc, [item.name]: item.id }),
-        {}
-      ),
-
-    /** @param {{ payload: DataSourceEntry }} action */
-    [types.create.success]: (state, action) => ({
-      ...state,
-      [action.payload.name]: action.payload.id,
-    }),
-
-    /** @param {{ payload: 'OK' | DataSource; meta: { id: string } }} action */
-    [types.postVersion.success]: (state, action) => {
-      if (action.payload === 'OK') return state;
-      return { ...state, [action.payload.name]: action.payload.id };
-    },
-  },
-});
-
-/** @type {VersionsCollection} */
-const initialVersions = {};
-const versions = createSlice({
-  initialState: initialVersions,
-  name: 'datasources/versions',
-  reducers: {},
-  extraReducers: {
-    /**
-     * @param {{ payload: DataSource[] }} action
-     * @returns {VersionsCollection}
-     */
-    [types.fetchAll.success]: (state, action) => {
-      const nextState = action.payload.reduce((acc, item) => {
-        if (state[item.name]) return acc;
-        return {
-          ...acc,
-          [item.name]: {
-            status: 'IDLE',
-            versions: [],
-          },
-        };
-      }, {});
-      return {
-        ...state,
-        ...nextState,
-      };
-    },
-    [types.fetchVersions.pending]: (state, action) => {
-      if (!action?.meta?.dataSource) return state;
-      return {
-        ...state,
-        [action.meta.dataSource]: {
-          status: 'PENDING',
-          versions: state[action.meta.dataSource].versions || [],
-        },
-      };
-    },
-    [types.fetchVersions.success]: (state, action) => {
-      if (!action?.meta?.dataSource) return state;
-      return {
-        ...state,
-        [action.meta.dataSource]: {
-          status: 'SUCCESS',
-          versions: action.payload,
-        },
-      };
-    },
-    [types.fetchVersions.fail]: (state, action) => {
-      if (!action?.meta?.dataSource) return state;
-      return {
-        ...state,
-        [action.meta.dataSource]: {
-          ...state[action.meta.dataSource],
-          status: 'FAIL',
-        },
-      };
-    },
-    [types.postVersion.success]: (state, action) => {
-      if (action.payload === 'OK') return state;
-      const {
-        payload: { name, id, versionDescription, versionId },
-      } = action;
-      return {
-        ...state,
-        [name]: {
-          ...state[name],
-          versions: state[name].versions.concat({
-            id,
-            versionDescription,
-            versionId,
-          }),
-        },
-      };
-    },
-  },
-});
 
 const dataSources = createSlice({
   initialState: entityAdapter.getInitialState(),
@@ -260,8 +121,7 @@ const error = (state = null, { type }) => {
 
 export const reducer = combineReducers({
   collection: dataSources.reducer,
-  activeVersions: activeVersions.reducer,
-  versions: versions.reducer,
+  versions: versionsReducer,
   status,
   error,
 });
@@ -274,10 +134,10 @@ export const selectors = {
     /** @param {State} state */
     state => state.dataSources.collection,
     /** @param {State} state */
-    state => state.dataSources.activeVersions,
-    (collectionState, activeVersionState) => {
+    state => state.dataSources.versions,
+    (collectionState, versions) => {
       const entities = baseSelectors.selectEntities(collectionState);
-      const activeIds = Object.values(activeVersionState);
+      const activeIds = Object.values(versions).map(item => item.active);
       return activeIds.map(id => entities[id]);
     }
   ),
